+++ dd-java-agent/tooling/src/main/java/datadog/trace/agent/tooling/HelperInjector.java
@@ -3,8 +3,9 @@ package datadog.trace.agent.tooling;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
 import java.util.Map;
 import java.util.Set;
 import lombok.extern.slf4j.Slf4j;
@@ -26,15 +27,17 @@ public class HelperInjector implements Transformer {
    * Construct HelperInjector.
    *
    * @param helperClassNames binary names of the helper classes to inject. These class names must be
-   *     resolvable by the classloader returned by DDAdvice#getAgentClassLoader()
+   *     resolvable by the classloader returned by DDAdvice#getAgentClassLoader(). Classes are
+   *     injected in the order provided. This is important if there is interdependency between
+   *     helper classes that requires them to be injected in a specific order.
    */
   public HelperInjector(final String... helperClassNames) {
-    this.helperClassNames = new HashSet<>(Arrays.asList(helperClassNames));
+    this.helperClassNames = new LinkedHashSet<>(Arrays.asList(helperClassNames));
   }
 
   private synchronized Map<TypeDescription, byte[]> getHelperMap() throws IOException {
     if (helperMap == null) {
-      helperMap = new HashMap<>(helperClassNames.size());
+      helperMap = new LinkedHashMap<>(helperClassNames.size());
       for (final String helperName : helperClassNames) {
         final ClassFileLocator locator =
             ClassFileLocator.ForClassLoader.of(Utils.getAgentClassLoader());
@@ -58,7 +61,29 @@ public class HelperInjector implements Transformer {
       synchronized (this) {
         if (!injectedClassLoaders.contains(classLoader)) {
           try {
-            new ClassInjector.UsingReflection(classLoader).inject(getHelperMap());
+            final Map<TypeDescription, byte[]> helperMap = getHelperMap();
+            final Set<String> existingClasses = new HashSet<>();
+            final ClassLoader systemCL = ClassLoader.getSystemClassLoader();
+            if (!classLoader.equals(systemCL)) {
+              // Build a list of existing helper classes.
+              for (final TypeDescription def : helperMap.keySet()) {
+                final String name = def.getName();
+                if (Utils.isClassLoaded(name, systemCL)) {
+                  existingClasses.add(name);
+                }
+              }
+            }
+            new ClassInjector.UsingReflection(classLoader).inject(helperMap);
+            if (!classLoader.equals(systemCL)) {
+              for (final TypeDescription def : helperMap.keySet()) {
+                // Ensure we didn't add any helper classes to the system CL.
+                final String name = def.getName();
+                if (!existingClasses.contains(name) && Utils.isClassLoaded(name, systemCL)) {
+                  throw new IllegalStateException(
+                      "Class was erroneously loaded on the System classloader: " + name);
+                }
+              }
+            }
           } catch (final Exception e) {
             log.error("Failed to inject helper classes into " + classLoader, e);
             throw new RuntimeException(e);
